fix(live): track recording duration as double for accurate display and subtitle sync
直播时长统计内部由 int 改为 double，渲染时再取整，消除逐分片/逐批 (int)
截断导致的显示偏差（已录制时长大于已刷新时长）。

同时利用更精确的累计时长修正字幕轴对齐：MpegtsTimestamp 改为先按 90kHz
缩放再取整，WebVttSub 偏移计算改用浮点除法，使字幕对齐从整秒级提升到亚秒级。

+++ src/N_m3u8DL-RE/DownloadManager/HTTPLiveRecordManager.cs
@@ -27,7 +27,7 @@ internal class HTTPLiveRecordManager
     DateTime? PublishDateTime;
     bool STOP_FLAG = false;
     bool READ_IFO = false;
-    ConcurrentDictionary<int, int> RecordingDurDic = new(); // 已录制时长
+    ConcurrentDictionary<int, double> RecordingDurDic = new(); // 已录制时长
     ConcurrentDictionary<int, double> RecordingSizeDic = new(); // 已录制大小
     CancellationTokenSource CancellationTokenSource = new(); // 取消Wait
     List<byte> InfoBuffer = new List<byte>(188 * 5000); // 5000个分包中解析信息，没有就算了
